Use categories reducer in categories components
