--- src/contexts/categories.context.jsx
@@ -1,27 +1,0 @@
-import { useState, createContext, useEffect } from 'react';
-import { getCategoriesAndDocuments } from '../utils/firebase/firebase.utils';
-
-export const CategoriesContext = createContext({
-  products: [],
-});
-
-export const CategoriesProvider = ({ children }) => {
-  const [categoriesMap, setCategoriesMap] = useState({});
-
-  const value = { categoriesMap, setCategoriesMap };
-
-  useEffect(() => {
-    const getCategoriesMap = async () => {
-      const categoryMap = await getCategoriesAndDocuments();
-      setCategoriesMap(categoryMap);
-    };
-
-    getCategoriesMap();
-  }, []);
-
-  return (
-    <CategoriesContext.Provider value={value}>
-      {children}
-    </CategoriesContext.Provider>
-  );
-};
