--- src/contexts/user.context.jsx
@@ -1,59 +1,0 @@
-import { createContext, useEffect, useReducer } from 'react';
-import {
-  onAuthStateChangedListener,
-  createUserDocumentFromAuth,
-} from '../utils/firebase/firebase.utils';
-import { createAction } from '../utils/reducer/reducer.utils';
-
-// actual value want to access
-export const UserContext = createContext({
-  currentUser: null,
-  setCurrentUser: () => null,
-});
-
-export const USER_ACTION_TYPES = {
-  SET_CURRENT_USER: 'SET_CURRENT_USER',
-};
-
-const userReducer = (state, action) => {
-  const { type, payload } = action;
-
-  switch (type) {
-    case USER_ACTION_TYPES.SET_CURRENT_USER:
-      return {
-        ...state,
-        currentUser: payload,
-      };
-    default:
-      throw new Error(`Unhandled type ${type}`);
-  }
-};
-
-const INITIAL_STATE = {
-  currentUser: null,
-};
-
-export const UserProvider = ({ children }) => {
-  const [state, dispatch] = useReducer(userReducer, INITIAL_STATE);
-
-  const { currentUser } = state;
-  const setCurrentUser = (user) => {
-    dispatch(createAction(USER_ACTION_TYPES.SET_CURRENT_USER, user));
-  };
-
-  const value = { currentUser, setCurrentUser };
-
-  useEffect(() => {
-    const unsubsribe = onAuthStateChangedListener((user) => {
-      if (user) {
-        createUserDocumentFromAuth(user);
-      }
-
-      setCurrentUser(user);
-    });
-
-    return unsubsribe;
-  }, []);
-
-  return <UserContext.Provider value={value}>{children}</UserContext.Provider>;
-};
