+++ src/routes/shop/shop.component.jsx
@@ -2,8 +2,23 @@ import { Routes, Route } from 'react-router-dom';
 import CategoriesPreview from '../categories-preview/categories-preview.component';
 import Category from '../category/category.component';
 import './shop.style.scss';
+import { getCategoriesAndDocuments } from '../../utils/firebase/firebase.utils';
+import { setCategoriesMap } from '../../store/categories/categories.action';
+import { useEffect } from 'react';
+import { useDispatch } from 'react-redux';
 
 const Shop = () => {
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    const getCategoriesMap = async () => {
+      const categoryMap = await getCategoriesAndDocuments();
+      dispatch(setCategoriesMap(categoryMap));
+    };
+
+    getCategoriesMap();
+  }, [dispatch]);
+
   return (
     <Routes>
       <Route index element={<CategoriesPreview />} />
