+++ src/store/categories/categories.selector.js
@@ -1,1 +1,1 @@
-export const selectCategories = (state) => state.categories.categoriesMap;
+export const selectCategoriesMap = (state) => state.categories.categoriesMap;
